k8s_sre_env/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "k8s-sre-env"
version = "0.1.0"
description = "K8s SRE Self-Improving RL Environment"
requires-python = ">=3.12"
dependencies = [
    "openenv-core==0.2.1",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0",
    "kubernetes>=28.0.0",
    "google-auth>=2.0.0",
    "requests>=2.25.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["k8s_sre_env*"]
